backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.3
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv
pytest>=7.0.0
httpx>=0.24.0
google-api-python-client>=2.0.0
googlemaps>=4.10.0
firebase-admin>=6.2.0
google-cloud-aiplatform>=1.35.0
slowapi>=0.1.9
cachetools>=5.3.0
asgiref>=3.7.2
